Removiendo componentes innecesarios
