+++ app/componentes/acciones.js
@@ -6,6 +6,7 @@ import { Grid, Row, Col, Button } from 'native-base';
 import { Pantalla } from './../pantalla';
 
 export default class Acciones extends Component {
+  
   ubicar(i){
     const m = Pantalla.margen, s = Pantalla.separacion / 2
     const primero = i == 0
